requirements.txt
aiogram==3.13.1
SQLAlchemy==2.0.36
aiosqlite==0.20.0
APScheduler==3.10.4
Pillow==10.4.0
python-dotenv==1.0.1
